+++ packages/cryptoplease/lib/app/screens/authenticated/investments/investments_screen.dart
@@ -2,7 +2,7 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../core/balances/presentation/refresh_balance_wrapper.dart';
-import '../../../../features/favorite_tokens/refresh_favorites.dart';
+import '../../../../features/favorite_tokens/widgets/extensions.dart';
 import '../../../../features/favorite_tokens/widgets/favorite_tokens_list.dart';
 import '../../../../features/investments/widgets/my_portfolio.dart';
 import '../../../../features/popular_tokens/widgets/extensions.dart';
