+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -7,10 +7,10 @@ import '../../../../core/currency.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/tokens/token_list.dart';
 import '../../../../di.dart';
-import '../../../../features/incoming_tip_payments/process_incoming_tip.dart';
-import '../../../../features/outgoing_direct_payments/build_context_ext.dart';
-import '../../../../features/outgoing_split_key_payments/module.dart';
-import '../../../../features/outgoing_tip_payments/module.dart';
+import '../../../../features/incoming_tip_payments/widgets/extensions.dart';
+import '../../../../features/outgoing_direct_payments/widgets/extensions.dart';
+import '../../../../features/outgoing_split_key_payments/widgets/extensions.dart';
+import '../../../../features/outgoing_tip_payments/widgets/extensions.dart';
 import '../../../../features/payment_request/models/payment_request.dart';
 import '../../../../features/qr_scanner/models/qr_scanner_request.dart';
 import '../../../../l10n/device_locale.dart';
