+++ packages/cryptoplease/lib/features/favorite_tokens/widgets/extensions.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'src/bl/bloc.dart';
+import '../src/bl/bloc.dart';
 
-extension RefreshFavorites on BuildContext {
+extension BuildContextExt on BuildContext {
   Future<void> refreshFavorites() => read<FavoritesBloc>().refresh();
 }
