+++ packages/cryptoplease/lib/features/incoming_tip_payments/widgets/extensions.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:uuid/uuid.dart';
 
-import '../../core/analytics/analytics_manager.dart';
-import '../../core/tip_payments.dart';
-import '../../di.dart';
-import '../../routes.gr.dart';
-import 'src/bl/it_bloc.dart';
-import 'src/widgets/link_listener.dart';
+import '../../../core/analytics/analytics_manager.dart';
+import '../../../core/tip_payments.dart';
+import '../../../di.dart';
+import '../../../routes.gr.dart';
+import '../src/bl/it_bloc.dart';
+import '../src/widgets/link_listener.dart';
 
 extension BuildContextExt on BuildContext {
   Future<void> processIncomingTip(TipPaymentData tipPayment) async {
