+++ packages/cryptoplease/lib/features/legal/flow.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../config.dart';
 import '../../core/presentation/utils.dart';
 
-extension LegalDocumentsFlowExt on BuildContext {
+extension BuildContextExt on BuildContext {
   void navigateToTermsOfUse() => openLink(termsUrl);
 
   void navigateToPrivacyPolicy() => openLink(privacyUrl);
