+++ packages/cryptoplease/lib/features/outgoing_direct_payments/widgets/extensions.dart
@@ -5,12 +5,12 @@ import 'package:provider/provider.dart';
 import 'package:solana/solana.dart';
 import 'package:uuid/uuid.dart';
 
-import '../../core/amount.dart';
-import '../../core/currency.dart';
-import '../../routes.gr.dart';
-import 'src/bl/bloc.dart';
+import '../../../core/amount.dart';
+import '../../../core/currency.dart';
+import '../../../routes.gr.dart';
+import '../src/bl/bloc.dart';
 
-extension OdpRoutesExt on BuildContext {
+extension BuildContextExt on BuildContext {
   void createAndOpenDirectPayment({
     required Decimal amountInUsdc,
     required Ed25519HDPublicKey receiver,
