+++ packages/cryptoplease/lib/features/outgoing_tip_payments/module.dart
@@ -2,12 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
-import 'package:uuid/uuid.dart';
 
 import '../../core/accounts/bl/account.dart';
 import '../../core/accounts/module.dart';
-import '../../core/amount.dart';
-import '../../core/analytics/analytics_manager.dart';
 import '../../di.dart';
 import 'src/bl/bloc.dart';
 import 'src/bl/repository.dart';
@@ -29,16 +26,4 @@ class OTModule extends SingleChildStatelessWidget {
         ],
         child: child,
       );
-}
-
-extension BuildContextExt on BuildContext {
-  String createOT({required CryptoAmount amount}) {
-    final id = const Uuid().v4();
-    final event = OTEvent.create(amount: amount, id: id);
-    read<OTBloc>().add(event);
-
-    sl<AnalyticsManager>().tipLinkCreated();
-
-    return id;
-  }
 }
